LayoutTests/media/track/track-webvtt-tc001-utf8.html

<html>
    <head>
        <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />

        <script src=../media-file.js></script>
        <script src=../video-test.js></script>
        <script>            

            var numberTests = 0;
            
            function track1Loaded()
            {
                findMediaElement();
                testExpected("video.textTracks[0].cues.length", "2");
                
                testExpected("video.textTracks[0].cues[0].id","1");
                testExpected("video.textTracks[0].cues[0].startTime","0.0");
                testExpected("video.textTracks[0].cues[0].endTime","30.5");
                testExpected("video.textTracks[0].cues[0].getCueAsSource()", "景気判断");
                
                testExpected("video.textTracks[0].cues[1].id","2");
                testExpected("video.textTracks[0].cues[1].startTime","31.0");
                testExpected("video.textTracks[0].cues[1].endTime","1200.5");
                testExpected("video.textTracks[0].cues[1].getCueAsSource()", "電力不足");

                allTestsEnded();
            }
            
            function track2Loaded()
            {
                findMediaElement();
                testExpected("video.textTracks[1].cues.length", "2");
                
                allTestsEnded();
            }
            
            function allTestsEnded()
            {
                numberTests++;
                if (numberTests >= 2)
                    endTest();
            }
        </script>
    </head>
    <body>
        <p>Tests that UTF-8 encoded characters are recognized properly and that different encodings (iconv) are not recognized as WebVTT a file (we do allow it, it just looks ugly).</p>
        <video>
            <track src="captions/tc001_utf8.vtt" onload="track1Loaded()">
            <track src="captions/tc001_iso2022jp3.vtt" onload="track2Loaded()">
        </video>
    </body>
</html>
