LayoutTests/media/track/track-webvtt-tc001-utf8-expected.txt

Tests that UTF-8 encoded characters are recognized properly and that different encodings (iconv) are not recognized as WebVTT a file (we do allow it, it just looks ugly).

EXPECTED (video.textTracks[0].cues.length == '2') OK
EXPECTED (video.textTracks[0].cues[0].id == '1') OK
EXPECTED (video.textTracks[0].cues[0].startTime == '0.0') OK
EXPECTED (video.textTracks[0].cues[0].endTime == '30.5') OK
EXPECTED (video.textTracks[0].cues[0].getCueAsSource() == '景気判断') OK
EXPECTED (video.textTracks[0].cues[1].id == '2') OK
EXPECTED (video.textTracks[0].cues[1].startTime == '31.0') OK
EXPECTED (video.textTracks[0].cues[1].endTime == '1200.5') OK
EXPECTED (video.textTracks[0].cues[1].getCueAsSource() == '電力不足') OK
EXPECTED (video.textTracks[1].cues.length == '2') OK
END OF TEST
